+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+
+## 0.2.0 - July 9, 2015
+
+### Added
+- Unit test for Jinja2 template handling
+- Mocking of other templates and macros
+- Test skipping
+
 ## 0.1.0 - March 9, 2015
 
 ### Added